backend/requirements.txt
fastapi==0.115.9
uvicorn
python-multipart
chromadb>=0.4.18
# langchain
requests
sse-starlette
aiofiles
python-magic
python-dotenv==1.0.0
pydantic[email]>=2.7.0,<3.0.0
pydantic-settings>=2.4.0,<3.0.0
pypdf==3.17.1
docx2txt==0.8
python-jose==3.3.0
passlib
bcrypt
sqlalchemy
databases[sqlite]